+++ packages/commons-ui-core/src/RichTypography/index.stories.js
@@ -1,5 +1,6 @@
-import { styled } from "@mui/material/styles";
 import React from "react";
+
+import Section from "../Section";
 
 import RichTypography from ".";
 
@@ -13,17 +14,6 @@ export default {
     },
   },
 };
-
-const Section = styled("div")(({ theme }) => ({
-  display: "flex",
-  [theme.breakpoints.up("sm")]: {
-    justifyContent: "center",
-    maxWidth: theme.contentWidths.values.sm,
-  },
-  [theme.breakpoints.up("md")]: {
-    maxWidth: theme.contentWidths.values.md,
-  },
-}));
 
 function Template({ content, ...args }) {
   return (
